[ENH] : 로그인 api 적용 및 connect next router 적용

+++ src/utils/apiManager.ts
@@ -1,7 +1,7 @@
-import axios from 'axios';
+import axios, { AxiosError } from 'axios';
 
 const instance = axios.create({
-  baseURL: 'https://api.dongjuind.co.kr/api/v1',
+  baseURL: 'https://api.dongjuind.co.kr/v1',
   timeout: 40000,
 
   // responseType:"json",
@@ -13,7 +13,7 @@ const instance = axios.create({
 });
 
 const formDataInstance = axios.create({
-  baseURL: 'https://api.dongjuind.co.kr/api/v1',
+  baseURL: 'https://api.dongjuind.co.kr/v1',
   timeout: 40000,
   headers: {
     'Content-Type': 'application/form-data',
@@ -23,13 +23,30 @@ const formDataInstance = axios.create({
 });
 
 const GET = async (entry: string, payload?: { params?: any }) => {
-  const data = await instance({
-    method: 'GET',
-    url: entry,
-    params: payload?.params,
-  });
+  if (typeof window !== 'undefined') {
+    try {
+      const data = await instance({
+        method: 'GET',
+        url: entry,
+        params: payload?.params,
+        headers: {
+          Authorization: localStorage.getItem('dongju-admin-token')
+            ? `${localStorage.getItem('dongju-admin-token')}`
+            : '',
+        },
+      });
 
-  return data.data;
+      return data.data;
+    } catch (error) {
+      const err = error as AxiosError;
+      const code = err.response?.status;
+      if (code === 403 || code === 401) {
+        alert('로그인 후 이용해 주세요');
+        window.location.href = '/login';
+      }
+      throw error;
+    }
+  }
 };
 
 const POST = async (
@@ -37,20 +54,42 @@ const POST = async (
   payload?: { bodyData: any },
   contentType?: 'form'
 ) => {
-  if (contentType === 'form') {
-    const data = await formDataInstance({
-      url: entry,
-      method: 'POST',
-      data: payload?.bodyData,
-    });
-    return data.data;
-  } else {
-    const data = await instance({
-      url: entry,
-      method: 'POST',
-      data: payload?.bodyData,
-    });
-    return data.data;
+  if (typeof window !== 'undefined') {
+    try {
+      if (contentType === 'form') {
+        const data = await formDataInstance({
+          url: entry,
+          method: 'POST',
+          data: payload?.bodyData,
+          headers: {
+            Authorization: localStorage.getItem('dongju-admin-token')
+              ? `${localStorage.getItem('dongju-admin-token')}`
+              : '',
+          },
+        });
+        return data.data;
+      } else {
+        const data = await instance({
+          url: entry,
+          method: 'POST',
+          data: payload?.bodyData,
+          headers: {
+            Authorization: localStorage.getItem('dongju-admin-token')
+              ? `${localStorage.getItem('dongju-admin-token')}`
+              : '',
+          },
+        });
+        return data.data;
+      }
+    } catch (error) {
+      const err = error as AxiosError;
+      const code = err.response?.status;
+      if (code === 403 || code === 401) {
+        alert('로그인 후 이용해 주세요');
+        window.location.href = '/login';
+      }
+      throw error;
+    }
   }
 };
 
@@ -58,22 +97,56 @@ const DELETE = async (
   entry: string,
   payload?: { params?: any; bodyData?: any }
 ) => {
-  const res = await instance({
-    url: entry,
-    method: 'DELETE',
-    params: payload?.params,
-    data: payload?.bodyData,
-  });
-  return res;
+  if (typeof window !== 'undefined') {
+    try {
+      const res = await instance({
+        url: entry,
+        method: 'DELETE',
+        params: payload?.params,
+        data: payload?.bodyData,
+        headers: {
+          Authorization: localStorage.getItem('dongju-admin-token')
+            ? `${localStorage.getItem('dongju-admin-token')}`
+            : '',
+        },
+      });
+      return res;
+    } catch (error) {
+      const err = error as AxiosError;
+      const code = err.response?.status;
+      if (code === 403 || code === 401) {
+        alert('로그인 후 이용해 주세요');
+        window.location.href = '/login';
+      }
+      throw error;
+    }
+  }
 };
 
 const PATCH = async (entry: string, { bodyData }: { bodyData: any }) => {
-  const res = await instance({
-    url: entry,
-    method: 'PATCH',
-    data: bodyData,
-  });
-  return res;
+  if (typeof window !== 'undefined') {
+    try {
+      const res = await instance({
+        url: entry,
+        method: 'PATCH',
+        data: bodyData,
+        headers: {
+          Authorization: localStorage.getItem('dongju-admin-token')
+            ? `${localStorage.getItem('dongju-admin-token')}`
+            : '',
+        },
+      });
+      return res;
+    } catch (error) {
+      const err = error as AxiosError;
+      const code = err.response?.status;
+      if (code === 403 || code === 401) {
+        alert('로그인 후 이용해 주세요');
+        window.location.href = '/login';
+      }
+      throw error;
+    }
+  }
 };
 
 export const API = { GET, POST, DELETE, PATCH };
